backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg[binary]==3.1.18
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
celery==5.3.6
redis==5.0.4

pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6
aiosqlite==0.20.0
